Breaking things
Kind of starting over, restructuring the program to create more abstraction. Through that process trying to figure out how to get methods in a subclass to take data from a text field and put them in an observableArrayList in a controller class which then updates a different arraylist in the main document controller class.